+++ apps/sv/src/main/scala/com/daml/network/sv/automation/SvDsoAutomationService.scala
@@ -319,7 +319,6 @@ class SvDsoAutomationService(
         triggerContext,
         dsoStore,
         participantAdminConnection,
-        config.acsCommitmentReconcilationInterval,
       )
     )
   }